[serve] Change stopping behavior (#43187)
Change from stop-fully-then-start to stop-then-immediately-start.
More concretely, previously we would make sure that the total number of (running or moving towards running) replicas **plus** the number of stopping replicas never exceeds the target. This PR changes it so that the number of stopping replicas is not considered in this calculation. The reasoning for this is that replicas that take a long time to gracefully shut down should not prevent new replicas from starting, as that will intentionally introduce lower availability.

Rollout plan:
* New behavior is enabled by default.
* Old behavior can be switched back on through the feature flag RAY_SERVE_STOP_FULLY_THEN_START_REPLICAS.
* Old behavior will be removed in the next release.

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>