[serve] Turn on `RAY_SERVE_ENABLE_QUEUE_LENGTH_CACHE` by default (#43169)
Turns on the queue length caching behavior by default. Leaves a few tests running with the flag off to verify that it works as expected in case we need to flip it back off.

While doing this I realized that streaming responses aren't supported over Ray Client, so I have a special case to disable this feature when using Ray Client to avoid a regression. Will follow up on this to properly support or disable in Ray 2.11.

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>