[doc] Fix color of button icon and text in dark mode button (#43267)
fixes the color of some icons and text on the ray-overview/use-cases page, which is currently unreadable in the dark theme.

This fix became slightly more complex because the icon was originally put in place using a pseudo-element, and you can't apply CSS styles that way. So instead, we improved the sphinx extension we use to create URL query parameters to make sure it parses content nested inside as usual (before it would treat any text nested inside as just text). This makes the query parameter extension machinery more useful and more robust.

---------

Signed-off-by: pdmurray <peynmurray@gmail.com>
Signed-off-by: Peyton Murray <peynmurray@gmail.com>
Co-authored-by: angelinalg <122562471+angelinalg@users.noreply.github.com>