[Data] Deprecate `Dataset.num_blocks()` for non-materialized `Dataset`s (#43178)
As part of the API simplification work in preparation for Ray Data GA, we are deprecating the `Dataset.num_blocks()` method. This method will only be available to `MaterializedDataset`s, and calling `Dataset.num_blocks()` on a non-materialized Dataset will result in a `NotImplementedError`.

Additional context behind the motivation for the change: We want to make `Block`s a Ray Data internal concept, so users should typically not need to be concerned with them. Instead, the primary method of choice should be `Dataset.count()`, which returns the number of rows in the Dataset.

The number of blocks is still available from method of the Dataset's internal `ExecutionPlan` object: `ds._plan.initial_num_blocks()`.

Signed-off-by: Scott Lee <sjl@anyscale.com>