[ci][wins/1] merge upload wheel logic into build container (#43421)
Context: this is a series of PR to move windows wheel to rayci (and deprecate .buildkite/windows_ci.sh). The mentioned script runs into issues when I try to upgrade it to python 3.11. Moving the script to rayci to uniform the window build environment.

Currently we use another container to upload the linux wheel. Merge the logic into the builder container, as a separated container is not necessary.

Signed-off-by: can <can@anyscale.com>