[data] Time backpressure (#43110)
Currently we are not timing how much time is spent in backpressure. This allows a per op timing for backpressure to be recorded, this can be used to inform how operators are being executed and how a set of backpressure policies are affecting the execution.

These stats are not propagated through to `DatasetStats` yet, they currently are only stored in `OpRuntimeMetrics`. Ideally we would be able to pass them along at a per operator level, but we could also pass along a total time spent in backpressure. A later PR will include this and other `StreamingExecutor` stats into the `DatasetStatsSummary`.

---------

Signed-off-by: Matthew Owen <mowen@anyscale.com>