[Dashboard] [Serve] Add application row with dropdown for deployment rows (#43506)
Adds back "application" rows to the top-level deployments page, which were removed in #42079. We keep the deployment rows as a dropdown under the application row, similar to how worker rows are a dropdown under node rows in the "Cluster" page (and we mimic the implementation from that part as well).

Per discussion offline:

Delete the "filters" component because it's not necessary for most use cases, and the user can always cmd-F to find in page if necessary
Start all the rows expanded
Replace the "deployments status" chip with "application status". If there's time, we can try to include both by making the metadata section number of columns configurable and set it to 4 instead of 3.
I also took a couple liberties which hadn't been discussed yet, but happy to revert them if they don't make sense:

Un-bold the the deployment name, but keep the application name bold. This highlights the hierarchical nature and makes the "deployment" rows more visually distinct from the "application" rows.
Make the number of replicas link to the deployment page where it shows the replicas. Reason: when I first saw "2" under replicas my first thought was "where can I see more about these 2 replicas", and I naturally wanted to click it.
---------

Signed-off-by: Archit Kulkarni <architkulkarni@users.noreply.github.com>
Co-authored-by: Huaiwei Sun <scottsun94@gmail.com>