[data] Sanitize ExecutionResources (#43474)
`ExecutionResources` serves as 2 different purposes: resource usage and resource limit. None values in it have different defaults (0 and inf) for those 2 use cases. This makes the class error prone, as users need to carefully handle None values. 

The best solution is to use 2 different classes for those 2 purposes. But it's already part of the public API. we don't want to break backward compatibility. 

This PR adds more wrapper methods to simplify the usage and avoid exposing None values to the external.  
This change will also remove many mypy typing hinting warnings of doing algebra operations between None and a float. 

---------

Signed-off-by: Hao Chen <chenh1024@gmail.com>