[core] If there's working_dir, don't set _py_driver_sys_path. (#43214)
In #31383 we add _py_driver_sys_path to all workers, making them search from the driver's path. This makes the workers able to find modules near the driver file. However, in these cases it's not working:

if there's working_dir. The code should search from working_dir, not from the driver's dir. e.g. lib code in driver's dir can update after job submission and before worker starts, making the runs inconsistent among workers.
if the worker is in a different node than the driver's node. The code would search "the driver's dir as in the node where the driver runs in", but in a different node. If there happen to be a stale code file in the worker node, the stale code is used.
(already handled) if it's dashboard, it's client mode, or it's interactive mode.
This PR addresses the first issue. Left a TODO for the second issue.

Part of #42863