[serve] Add `ReplicaID` class (#43557)
Unifies replica IDs on a ReplicaID class (similar to and including DeploymentID). This is used to uniquely identify replicas and carry around their metadata.

This allows us to clean up the strings in some of our logs (where the replica_id contained redundant information about the application and deployment name).

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>
Signed-off-by: David Nie <1193115470@qq.com>