[Data] Fix flaky data tests (#43726)
This hopefully fixes `test_streaming_integration.py` which has been consistently failing and was moved to flaky. The test suite seems to be sensitive to ordering of the tests specifically on the CI machines. This moves a test to the bottom which appears to solve the issue (tested several times directly on a CI machine) and also in a manually triggered [postmerge test run](https://buildkite.com/ray-project/postmerge/builds/3331#018e1595-209f-4028-a8d4-40fa393b095d). More investigation can be done later, but for now applying this bandaid.

Signed-off-by: Matthew Owen <mowen@anyscale.com>
Signed-off-by: David Nie <1193115470@qq.com>