[Serve] Fix deployment route prefix override by default route prefix from serve run cli (#43805)
The default route_prefix ("/") on serve run cli is wrongly overrode the route_prefix from the deployment. If the route_prefix is not passed into the serve run cli, we should keep route_prefix as DEFAULT.VALUE, so the downstream serve.run api can determine how to override. Also added a test to ensure this will be caught in the future.

---------

Signed-off-by: Gene Su <e870252314@gmail.com>