[serve] active node compaction (#466)
[serve] active node compaction

Implement active node compaction.

Deployment scheduler:
* `detect_compact_opportunities` looks for a node that can become idle
if all its replicas are moved to other nodes in the cluster. It chooses
the node with the least number of replicas to migrate. If it identifies
a node that can be compacted, it will update state
`self._active_compaction`.
* The current active compaction will complete once the target node
becomes idle (there are no more replicas running on it).
* The current active compaction will be cancelled if new replicas have
been scheduled onto that node *since* the compaction opportunity was
first identified. We'd expect this to happen if upscaling happened in
the middle of compaction, and we take this as a sign that the compaction
is unable to complete.

Deployment state:
* When an active compaction is *cancelled*, any replicas that are still
`PENDING_MIGRATION` on that node will return to `RUNNING` to avoid edge
cases where we might induce dropped availability.