[serve] defensively guard against scheduling edge case (#43839)
Theoretically this can happen:

In the deployment state manager / deployment scheduler we still think a replica is running on a node but ray.nodes() no longer returns any info on that node. To guard against this potential race condition (when info is out of sync), use .get() instead of directly accessing dictionaries.

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>
Signed-off-by: David Nie <1193115470@qq.com>