[CI] [Lint] [Dashboard] Make frontend linter automatically fix lint issues (#43515)
When making dashboard frontend changes, the linter outputs things like

`  0 errors and 7 warnings potentially fixable with the `--fix` option.`

This PR adds the `--fix` command to the linter script to automatically fix lint issues when possible.  This is consistent with how the other linters work in the script for non-frontend code.

Signed-off-by: Archit Kulkarni <architkulkarni@users.noreply.github.com>
Signed-off-by: David Nie <1193115470@qq.com>