[Serve] Disallow setting max_replicas_per_node and placement_group_bundles options together (#43872)
The semantic of max_replicas_per_node with non-strict-pack PG is not well defined so we disallow setting them together. While strict-pack PG should work with max_replicas_per_node but currently there is no use case so we disallow that as well. In the future, we can support strict-pack PG with max_replicas_per_node if needed.

Signed-off-by: Jiajun Yao <jeromeyjj@gmail.com>