[Core] Ignore worker lease requests from dead caller (#43920)
Currently we clear all the pending lease requests when raylet receives the worker death notification. However if the lease request comes in after the notification (CoreWorker calls Disconnect and then Shutdown, the worker death notification happen after Disconnect but at that time there can be in-flight lease requests to the raylet), the lease request won't have chance to be cleared. This PR fixes this by ignoring those lease requests.

Signed-off-by: Jiajun Yao <jeromeyjj@gmail.com>