[Cluster launcher] [Azure] Allow configurable managed identity when launching Azure clusters (#43897)
Currently, one is not able to use a previously configured managed identity when launching Azure clusters. This can cause issues with trying to connect multiple clusters to shared resources such as storage accounts or key vaults that lie outside the resource group of the created VMs as they won't able to access those shared resources. It also makes it much easier to tear down clusters, as the entire resource group can be deleted without fear of deleting any shared resources between clusters.

This change allows the user to pass the name of an existing managed identity and its resource group as part of the provider options in the cluster yaml file. If those fields are provided, a new managed identity is not created and the existing one is used. This change is backwards-compatible since if the new fields are not provided, the default behavior is the same as before - create a new managed identity in the same resource group as the deployment. I've verified both cases locally by launching a cluster with/without these parameters specified and viewing the (created) MSIs in the portal.

Related issue number
Discussed in #29215, but does not appear that it was ever PRed.

---------

Signed-off-by: Mike Danielczuk <michael.danielczuk@ambirobotics.com>