[Serve] Fix serve deploy not getting current directory added to sys.path (#43899) (#44011)
Cherry-pick #43899

Re: #43214 There is a core change that stopped adding script directory and current directory for different conditions. This accidentally stopped adding current directory when it's coming from the dashboard where in that case we only want to stop adding the script directory. This PR fixes the issue and allow serve deploy to continue working with the current directory.

---------

Signed-off-by: Gene Su <e870252314@gmail.com>