[serve] Drop stale handle metrics  (#44045)
We don't currently drop handle metrics after a handle has been removed (i.e., a replica or proxy with a handle has been shut down). This causes "phantom" metrics to prevent downscaling.

This PR adds a timeout after which we drop the metrics: 2 * the reporting period (with a constant floor of 10s in case the reporting period is configured to be very low).

Also cleans up a few log statements.

---------

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>
Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>
Co-authored-by: Cindy Zhang <cindyzyx9@gmail.com>
Co-authored-by: shrekris-anyscale <92341594+shrekris-anyscale@users.noreply.github.com>
Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>