[core] Streaming generator task waits for all object report acks before finishing the task (#44079)
This fixes a bug in handling task (worker process) failures for streaming generator tasks. Streaming generator executors report the values of the dynamic returns to the caller asynchronously during the task execution. If a report is lost, then the caller will not be able to get the value of the yield'ed ObjectRef. Therefore, we need to wait until all in-flight object reports have been acked before we can finish the streaming generator task. Otherwise, it is possible for the caller to think that the task has finished, but then the executor dies before it can report the dynamic return object's value.

Also cleans up some generator_waiter.cc headers, etc.

Signed-off-by: Stephanie Wang <swang@cs.berkeley.edu>