[ci][release-automation] Add docker group permission for forge (#44363)
Add Docker group for forge. The reason is that step to check Ray commit in nightly image is failing on master: https://buildkite.com/ray-project/release-automation/builds/313 since I only tested it when docker group was added in forge_x86_64.Dockerfile. The group was removed before merge which causes the build to not have Docker permission.

Also add some logs.

Signed-off-by: khluu <51931015+khluu@users.noreply.github.com>