Fixes the flaky mutable_object_test
Based on a local issue I resolved, I believe this commit should fix the
flaky mutable_object_test in #44396.

The header of the mutable object was not being explicitly initialized.
In some cases, this caused deadlock due to a spinlock backed by garbage memory.

This commit explicitly initializes the header, which should resolve the
issue.

Tested: mutable_object_test

Signed-off-by: Jack Humphries <1645405+jackhumphries@users.noreply.github.com>