[Doc] Make runtimeEnvYAML more prominent in docs (#44326)
When using Ray Jobs, it's better to specify the runtime_env in the Job spec, rather than in ray.init() in the entrypoint script. The reason is that it's simpler; in the first case the entrypoint script runs in the runtime env (as well as any tasks/actors), whereas in the second case only the tasks and actors run in the runtime env.

Some users are still using runtime_env in ray.init(). One possible reason is that it's not prominent enough in the docs. This PR adds a reference to KubeRay RayJob in the Ray Core runtime env docs, and also adds a code snippet to make it more friendly for users to copy and paste.

---------

Signed-off-by: Archit Kulkarni <architkulkarni@users.noreply.github.com>
Co-authored-by: angelinalg <122562471+angelinalg@users.noreply.github.com>