[core] Support spill objects to node_id sub-directory (#44341)
When system-config was used to specify the spilling location, only one direction can be used which is the same for all the workers. When the dir is a shared NFS, conflict might happen.

This PR adds support for using <NODE_ID> as part of the spilling directory name to avoid the conflicts. Specifically,
the spilling directory would be set to /tmp/spill/ray_spilled_objects_<NODE_ID>. Note that this change is not backwards compatible since it changes the spill directory name. We preferred this as opposed to introducing an additional config argument (API change) that controls whether to use <NODE_ID> as part of the directory name.

Signed-off-by: Rui Qiao <ruisearch42@gmail.com>