[Data] Change Parquet encoding ratio lower bound from 2 to 1 (#44470)
Ray Data estimates the in-memory size of file-based datasets. To estimate the size of Parquet datasets, Ray Data samples rows and estimates the ratio between in-memory data size and raw data size. If this ratio is off, Ray Data can perform too few block splits.

Currently, Ray Data clamps the estimated ratio to 2. However, for some Parquet datasets, the actual ratio is closer to 1. So, this PR lowers to lower bound from 2 to 1.

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>