[core] Deflake test_actor_client_mode (#44471)
test_actor_client_mode is flaky on CI Mac boxes. It was observed that later spills are blocked because the 1st spill of 800M has not finished, therefore a timeout. This change is to reduce the object size and memory size by 10X so that the spill can finish in much shorter time.

Signed-off-by: Rui Qiao <ruisearch42@gmail.com>