[Data] Fix bug where debug logs are logged to console (#44601)
The "ray" parent logger prints logs to console, and "ray.data" logger propagates logs to the "ray" parent logger. The effective level of the "ray" logger is INFO, so I expected DEBUG logs to get rejected. However, the effective level of a logger is only considered when a log is first emitted, and not when it's passed to parent loggers. So, DEBUG logs are passed to the "ray" handlers, and DEBUG logs get printed to console.

To fix this issue, this PR disables propagation to the "ray" logger and independently configures logging to console.

---------

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>