[core] Expose python type ray.ClusterID, and use it in to track actor and task last export time. (#44490)
We had a field _last_export_session_and_job in class RemoteFunction and class _ActorClassMetadata. It's used to track whether this actor/task has been exported to GCS or not, without having to pay an RTT to GCS for internal_kv_exists. The problem is this approach is not robust: the "session" is not a real concept, it's just an incrementing int counter in the class; while the Job IDs are not unique among clusters - all clusters' first job are 01000000. This requires the field to be omitted in pickling which makes unnecessary exporting to GCS.

This PR uses (cluster_id, job_id) to track last export time, because the cluster ID is unique. To achieve this, we need to also expose the ClusterID from a pure C++ concept to also be a Python type.

Signed-off-by: Ruiyang Wang <rywang014@gmail.com>