[Jobs] [Logs] Stream runtime env log to job log file (#44405)
When a job is submitted, it stays in a "PENDING" state before becoming "RUNNING". In practice, this is nearly always because the runtime_env is being installed. However, it's not obvious where to see the progress of the runtime_env installation.

For better observability, this PR streams the runtime_env log to the existing job-driver-<submission_id>.log file.

Note that currently the Job log API itself won't return the runtime_env setup log until the job starts, because the job_head dashboard module needs to know which node the runtime env is being installed on before the job starts, and we don't have a way of surfacing that information yet. I'll file an issue and link it here. The log file itself will still have the runtime_env setup logs written to it in real time.

Currently this feature is gated behind an environment variable (defaults to OFF), in case the logs are too spammy.

---------

Signed-off-by: Archit Kulkarni <architkulkarni@users.noreply.github.com>