[Data] Add ability to not print certain logs to console (#44661)
All info, warning, and error logs are printed to console. However, we don't want to print some errors to the console (specifically, we don't want to print full tracebacks from some UDF errors). To avoid printing these errors, we use DEBUG logs rather than WARNING logs. The issue with this approach is that it makes it harder to configuring logging for ERROR level logs.

This PR solves this issue by adding a flag to not write logs to console.

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>