[core] Adds ActorUnavailableError. (#44212)
When an actor process died, the ray.get caller gets an RayActorError, signaling the actor's death. When the actor is not dead but in a temporal unavailable state, however, we still raised that error. This makes user code hard to handle the case where the actor recovered. Example cases:

The actor died, but it set max_restarts and is currently in RESTARTING state (e.g. running init).
The actor suffers from a temporal unstable network issue, e.g. connection broke but can re-establish.
This PR introduces a new ActorUnavailableError, indicating that situation. When an actor is temporarily unavailable, a ray.get raises that error. Note that this makes no assumptions on the side effects - the actor may have well observed the side effect and prepared the return value, just it can't reply back. Also it does not guarantee the actor will eventually recover - think a death loop in init exhausting max_restarts, or a unstable network finally turned to a blackout.

To make the changes backward compatible, Now we make RayActorError as a base class with 2 subclasses: the ActorUnavailableError as mentioned before, and ActorDiedError for the death cases. Both subclasses share the same set of attributes: actor_id: str, error_msg: str, actor_init_failed: bool, preempted: bool.

Streaming generator behavior: In the middle of a streaming process, if the actor becomes unavailable, a next(gen) call will get StopIteration, i.e. early stop. A ray.get(obj_ref) on the last received object_ref will get ActorUnavailableError.

Signed-off-by: Ruiyang Wang <rywang014@gmail.com>
Signed-off-by: Ruiyang Wang <56065503+rynewang@users.noreply.github.com>
Co-authored-by: Stephanie Wang <swang@cs.berkeley.edu>
Co-authored-by: shrekris-anyscale <92341594+shrekris-anyscale@users.noreply.github.com>
Co-authored-by: angelinalg <122562471+angelinalg@users.noreply.github.com>