[Core] Fix flaky mutable object test (#44720)
mutable_object_test is currently "flaky" because the test sometimes times out in CI. This happens because there is significant contention on the header_sem semaphore in PlasmaObjectHeader::ReadAcquire() when there are many readers, due to polling.

As a temporary fix, I have the threads yield in each iteration of the loop. The correct long-term solution is to use a futex (or something similar) to avoid polling, which I will implement soon. But I want to merge this PR first to avoid blocking submission for others.

Signed-off-by: Jack Humphries <1645405+jackhumphries@users.noreply.github.com>