[ci] run ci state machine with shorter data length (#44783)
Currently the CI state machine uses a very long historical data to make judgment (100 tests). As we're moving things to run periodically, adjust the setting to work with a shorter history. I also adjust flaky percentage to 7% (so that it's 2 red per 30 run, which is 6.6%, is not punished, but 3 red per 30 run is)

Signed-off-by: can <can@anyscale.com>