[Serve] fix getting attributes on stdout during Serve logging redirect (#44787)
There are attributes on sys.stdout and sys.stderr which are not implemented on our StreamToLogger class. When we overrided them for redirect, it will raise AttributeError. This caused Ray Tune unable to get the attributes on those objects on Anyscale Services where the redirect is turned on. This PR added __getattr__() on the StreamToLogger to get the attributes on the original object so they can still be accessible. Also added a unit test for it.

---------

Signed-off-by: Gene Su <e870252314@gmail.com>