[dashboard] Fix dashboard stack trace error when sudo is not found (#43857)
When Ray runs on k8s with the images that does not install "sudo" and runs as non-root in containers, the "Stack Trace (Driver)" of the dashboard goes wrong and returns "sudo FileNotFound". In fact, py-spy can be used even if sudo is not installed.
The code previous forget to handle the case where the sudo FileNotFound exception is thrown and directly returns it to the user. This PR fixes this issue, treat sudo not found as "cannot do passwordless sudo"