[dashboard] Fixes the CPU percent metrics for Dashboard process.
The metrics `ray_component_cpu_percentage` for Dashboard is always 0. This is because every time we create a fresh `psutil.Process` instance. In the document it says:

    When *interval* is 0.0 or None (default) compares process times
    to system CPU times elapsed since last call, returning
    immediately (non-blocking). That means that the first time
    this is called it will return a meaningful 0.0 value.

... so we always get a 0.0 value. Further, in Kubernetes the CPU percentage should be based on the amount of CPUs available to the container. Both issues are already fixed in the `reporter_agent.py` file. We need to make the dashboard metrics comparable to that.

This PR moves the CPU percentage calculation to a shared function in `metrics_utils.py` and uses it in both the agent code and the head code.

Signed-off-by: Ruiyang Wang <rywang014@gmail.com>