[core][experimental] Pass torch.Tensors through accelerated DAGs (#44825)
This PR adds support for passing torch.Tensors to local actors in an
accelerated DAG, via Ray's shared memory store. It supports the
following transfer cases, as long as the sending and receiving actors
are on the same node: CPU-CPU, CPU-GPU, GPU-CPU, GPU-GPU (via CPU).

This iteration requires the user to explicitly declare which DAG nodes
contain torch.Tensors and the tensors' shape and dtype, with a new
`with_type_hint` decorator. For example:

```python
    with InputNode() as inp:
        dag = sender.send.bind(inp)
        dag = dag.with_type_hint(TorchTensorType(SHAPE, DTYPE))
        dag = receiver.recv.bind(dag)

    compiled_dag = dag.experimental_compile()
```
This declaration isn't necessarily useful for this PR, but it is
included now because it makes it much simpler to efficiently support
other cases in the future, such as p2p GPU-GPU transfers.

When a TorchTensor node is declared, the serialization of the underlying
torch.Tensor is performed differently from vanilla Ray. In particular,
we store the numpy view of the data. On the receiving actor, we
deserialize to a torch.Tensor and move it to the device assigned to the
actor, if any. Microbenchmarking shows that this is 4x faster than
normal pickling and unpickling of a torch.Tensor, likely due to Ray's
serialization support for numpy. Also, when moving the torch.Tensor to a
GPU on the receiving side, we can avoid one extra data copy by copying
directly from Ray's shared memory buffer to GPU memory.

Limitations:
- Only supports tasks that directly return a torch.Tensor, i.e. the
torch.Tensor cannot be nested in other data.
- The task must declare the shape and dtype of its torch.Tensor at DAG
compile time.
- Does not support local p2p GPU-GPU transfer, either using `cudaMemCpy`
or NCCL. Microbenchmark shows this can be >10x faster than transfer via
CPU.
- Does not support multinode GPU-GPU transfer, e.g., via RPC between
hosts or NCCL.

---------

Signed-off-by: Stephanie Wang <swang@cs.berkeley.edu>